Split IRenderer implementer classes out to their own files.
